@@ -1,42 +1,40 @@
 {
-  "_args": [
-    [
-      "ms@2.1.2",
-      "D:\\Angular\\gatito-book\\gatitobook\\api"
-    ]
-  ],
-  "_from": "ms@2.1.2",
-  "_id": "ms@2.1.2",
+  "_from": "ms@^2.1.1",
+  "_id": "ms@2.1.3",
   "_inBundle": false,
-  "_integrity": "sha512-sGkPx+VjMtmA6MX27oA4FBFELFCZZ4S4XqeGOXCv68tT+jb3vk/RyaKWP0PTKyWtmLSM0b+adUTEvbs1PEaH2w==",
+  "_integrity": "sha512-6FlzubTLZG3J2a/NVCAleEhjzq5oxgHyaCU9yYXvcLsvoVaHJq/s5xXI6/XXP6tz7R9xAOtHnSO/tXtF3WRTlA==",
   "_location": "/needle/ms",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "ms@2.1.2",
+    "raw": "ms@^2.1.1",
     "name": "ms",
     "escapedName": "ms",
-    "rawSpec": "2.1.2",
+    "rawSpec": "^2.1.1",
     "saveSpec": null,
-    "fetchSpec": "2.1.2"
+    "fetchSpec": "^2.1.1"
   },
   "_requiredBy": [
     "/needle/debug"
   ],
-  "_resolved": "https://registry.npmjs.org/ms/-/ms-2.1.2.tgz",
-  "_spec": "2.1.2",
-  "_where": "D:\\Angular\\gatito-book\\gatitobook\\api",
+  "_resolved": "https://registry.npmjs.org/ms/-/ms-2.1.3.tgz",
+  "_shasum": "574c8138ce1d2b5861f0b44579dbadd60c6615b2",
+  "_spec": "ms@^2.1.1",
+  "_where": "D:\\3ADigitall\\api\\node_modules\\needle\\node_modules\\debug",
   "bugs": {
-    "url": "https://github.com/zeit/ms/issues"
+    "url": "https://github.com/vercel/ms/issues"
   },
+  "bundleDependencies": false,
+  "deprecated": false,
   "description": "Tiny millisecond conversion utility",
   "devDependencies": {
-    "eslint": "4.12.1",
+    "eslint": "4.18.2",
     "expect.js": "0.3.1",
     "husky": "0.14.3",
     "lint-staged": "5.0.0",
-    "mocha": "4.0.1"
+    "mocha": "4.0.1",
+    "prettier": "2.0.5"
   },
   "eslintConfig": {
     "extends": "eslint:recommended",
@@ -48,7 +46,7 @@
   "files": [
     "index.js"
   ],
-  "homepage": "https://github.com/zeit/ms#readme",
+  "homepage": "https://github.com/vercel/ms#readme",
   "license": "MIT",
   "lint-staged": {
     "*.js": [
@@ -61,12 +59,12 @@
   "name": "ms",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/zeit/ms.git"
+    "url": "git+https://github.com/vercel/ms.git"
   },
   "scripts": {
     "lint": "eslint lib/* bin/*",
     "precommit": "lint-staged",
     "test": "mocha tests.js"
   },
-  "version": "2.1.2"
+  "version": "2.1.3"
 }
