@@ -7,7 +7,7 @@ import { LoginComponent } from './login/login.component';
 import { CadastroComponent } from './cadastro/cadastro.component';
 import { HomeComponent } from './home/home.component';
 import { EsqueciSenhaComponent } from './esqueci-senha/esqueci-senha.component';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { MensagemComponent } from './mensagem/mensagem.component';
 
@@ -18,13 +18,14 @@ import { MensagemComponent } from './mensagem/mensagem.component';
     CadastroComponent,
     HomeComponent,
     EsqueciSenhaComponent,
-    MensagemComponent,
+    MensagemComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     FormsModule,
-    HttpClientModule
+    HttpClientModule,
+    ReactiveFormsModule,
   ],
   exports: [ MensagemComponent ],
   providers: [],
